+++ BallancingRobot_V1/Core/Src/motor_simple.c
@@ -8,7 +8,7 @@
 #include "main.h"
 #include "motor_simple.h"
 
-MotorStatus_t Motor_SetRideParameters(Motor_t *Motor, uint8_t PWM, uint8_t Dir)
+MotorStatus_t Motor_SetRideParameters(Motor_t *Motor, uint16_t PWM, uint16_t Dir)
 {
 	//Make sure PWM is in range from 0 to 100
 	if (PWM > MOTOR_PWM_MAX)
